requirements.txt
psycopg2-binary
asgiref==3.8.1
django==5.2.1
djangorestframework
environs
requests
pip==25.1.1
python-dotenv
setuptools==80.7.1
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
